services/ai-server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
